Updated views and some tweeks
