+++ Projects/Claims/Sources/Views/SubmitClaim/Pests/SubmitClaimPestsScreen.swift
@@ -31,15 +31,15 @@ struct SubmitClaimPestsScreen: View {
                 .sectionContainerStyle(.transparent)
 
                 VStack(spacing: 4) {
-                    infoExpandableView(
+                    InfoExpandableView(
                         title: L10n.submitClaimWhatCostTitle,
                         text: L10n.submitClaimGlassDamageWhatCostLabel
                     )
-                    infoExpandableView(
+                    InfoExpandableView(
                         title: L10n.submitClaimHowBookTitle,
                         text: L10n.submitClaimGlassDamageHowBookLabel
                     )
-                    infoExpandableView(
+                    InfoExpandableView(
                         title: L10n.submitClaimWorkshopTitle,
                         text: L10n.submitClaimGlassDamageWorkshopLabel
                     )
@@ -47,43 +47,7 @@ struct SubmitClaimPestsScreen: View {
                 .padding(.vertical, 8)
 
                 SupportView()
-                    .padding(.vertical, 32)
-            }
-        }
-    }
-
-    func infoExpandableView(title: String, text: String) -> some View {
-        hSection {
-            hRow {
-                hText(title)
-                    .lineLimit(1)
-            }
-            .withCustomAccessory({
-                Spacer()
-                Image(
-                    uiImage: selectedFields.contains(title)
-                        ? hCoreUIAssets.minusSmall.image : hCoreUIAssets.plusSmall.image
-                )
-                .transition(.opacity.animation(.easeOut))
-            })
-            .onTap {
-                if !selectedFields.contains(title) {
-                    selectedFields.append(title)
-                } else {
-                    if let index = selectedFields.firstIndex(of: title) {
-                        selectedFields.remove(at: index)
-                    }
-                }
-            }
-            .hWithoutDivider
-
-            if selectedFields.contains(title) {
-                VStack(alignment: .leading) {
-                    hRow {
-                        hText(text)
-                            .foregroundColor(hTextColorNew.secondary)
-                    }
-                }
+                    .padding(.vertical, 56)
             }
         }
     }
@@ -91,6 +55,7 @@ struct SubmitClaimPestsScreen: View {
 
 struct SubmitClaimPestsScreen_Previews: PreviewProvider {
     static var previews: some View {
-        SubmitClaimPestsScreen()
+        Localization.Locale.currentLocale = .en_SE
+        return SubmitClaimPestsScreen()
     }
 }
